+++ tests/bun/tests/server-no-cache.test.ts
@@ -57,7 +57,7 @@ test("it does not cache and returns webp", async () => {
   expect(res.headers.get("Content-Type")).toBe("image/webp");
 
   const cachedFile = Bun.file(
-    "./data/images/cat-png-w-100-h-100-fit-base.webp",
+    "./data/images/public/cat-png-w-100-h-100-fit-base.webp",
   );
   expect(await cachedFile.exists()).toBe(false);
 });
@@ -68,7 +68,7 @@ test("it does not cache and returns avif", async () => {
   expect(res.headers.get("Content-Type")).toBe("image/avif");
 
   const cachedFile = Bun.file(
-    "./data/images/cat-png-w-100-h-100-fit-base.avif",
+    "./data/images/public/cat-png-w-100-h-100-fit-base.avif",
   );
   expect(await cachedFile.exists()).toBe(false);
 });
@@ -77,6 +77,8 @@ test("it does not cache and returns original format", async () => {
   const res = await fetch(origin + "?src=/cat.png&w=100&h=100");
   expect(res.status).toBe(200);
 
-  const cachedFile = Bun.file("./data/images/cat-png-w-100-h-100-fit-base.png");
+  const cachedFile = Bun.file(
+    "./data/images/public/cat-png-w-100-h-100-fit-base.png",
+  );
   expect(await cachedFile.exists()).toBe(false);
 });